Cambios luis

+++ src/main/java/com/bew/demo/dao/DictamenRepository.java
@@ -10,12 +10,13 @@ import org.springframework.data.repository.query.Param;
 import com.bew.demo.model.Dictamen;
 
 
-public interface DictamenRepository  extends JpaRepository <Dictamen,Integer>{
+public interface DictamenRepository  extends JpaRepository <Dictamen,Long>{
 	
 	@Query(
             value = "SELECT s FROM Dictamen s WHERE s.idAlumno = :idAlumno",
             nativeQuery = false)
-    Optional<Dictamen> findByIdAlumno(@Param("idAlumno") Integer idAlumno);
+    Optional<Dictamen> findByIdAlumno(@Param("idAlumno") Long idAlumno
+    		);
 	
 	@Query(
             value = "SELECT s FROM Dictamen s WHERE s.estado = :estado",
