+++ src/main/java/com/bew/demo/dao/LiberacionExtempRepository.java
@@ -11,12 +11,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 @Repository
-public interface LiberacionExtempRepository extends JpaRepository <LiberacionExtemp,Integer>{
+public interface LiberacionExtempRepository extends JpaRepository <LiberacionExtemp,Long>{
 	
 	@Query(
             value = "SELECT s FROM LiberacionExtemp s WHERE s.idAlumno = :idAlumno",
             nativeQuery = false)
-    Optional<LiberacionExtemp> findByIdAlumno(@Param("idAlumno") Integer idAlumno);
+    Optional<LiberacionExtemp> findByIdAlumno(@Param("idAlumno") Long idAlumno);
 	
 	@Query(
             value = "SELECT s FROM LiberacionExtemp s WHERE s.estado = :estado",
