+++ src/main/java/com/bew/demo/dto/ClienteDTO.java
@@ -1,6 +1,10 @@
 package com.bew.demo.dto;
 
+import lombok.Getter;
+import lombok.Setter;
+
 import java.io.Serializable;
+
 
 public class ClienteDTO implements Serializable{
 	
@@ -8,19 +12,21 @@ public class ClienteDTO implements Serializable{
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private Integer idCliente;
+	private Long idCliente;
 	private String nombre;
-	
-	
-	public Integer getIdCliente() {
+
+	public Long getIdCliente() {
 		return idCliente;
 	}
-	public void setIdCliente(Integer idCliente) {
+
+	public void setIdCliente(Long idCliente) {
 		this.idCliente = idCliente;
 	}
+
 	public String getNombre() {
 		return nombre;
 	}
+
 	public void setNombre(String nombre) {
 		this.nombre = nombre;
 	}
